+++ app/src/types/index.ts
@@ -1,3 +1,6 @@
 export * from './toast-type'
 export * from './tips-event-type'
-export * from './types'
+export * from './member'
+export * from './team'
+export * from './inputs/team-input'
+export * from './inputs/member-input'
