+++ app/src/types/inputs/member-input.ts
@@ -1,0 +1,6 @@
+export interface MemberInput {
+  name: string
+  walletAddress: string
+  id?: string
+  isValid: boolean
+}
