+++ app/src/types/inputs/team-input.ts
@@ -1,0 +1,7 @@
+import type { MemberInput } from "./member-input"
+
+export interface TeamInput {
+  name: string
+  description: string
+  members: MemberInput[]
+}
