+++ app/src/types/member.ts
@@ -1,0 +1,6 @@
+export interface Member {
+  id: string
+  name: string
+  walletAddress: string
+  teamId: number
+}
