+++ app/src/types/team.ts
@@ -1,0 +1,8 @@
+import type { Member } from "./member"
+
+export interface Team {
+  id: string
+  name: string
+  description: string
+  members: Member[]
+}
