+++ app/src/views/TeamView.vue
@@ -30,10 +30,9 @@ import { onMounted, ref } from 'vue'
 import { useRouter } from 'vue-router'
 import { FetchTeamAPI } from '@/apis/teamApi'
 const router = useRouter()
-import { type TeamInput, type Team } from '@/types/index'
+import { ToastType, type Team, type TeamInput } from '@/types'
 import { isAddress } from 'ethers' // ethers v6
 import { useToastStore } from '@/stores/toast'
-import { ToastType } from '@/types'
 
 const { show } = useToastStore()
 const teamApi = new FetchTeamAPI()
